refactor: create power_adapters.py to decouple stats/ from db access and move tests
